% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_atc_table.R
\name{build_atc_tbl}
\alias{build_atc_tbl}
\title{Build an ATC table from page data}
\usage{
build_atc_tbl(pagedata)
}
\arguments{
\item{pagedata}{Page data, created through extract_atc_hierarchy.}
}
\value{
An ATC table.
}
\description{
Build an ATC table from page data
}
