% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_atc_table.R
\name{clean_DDD_cols}
\alias{clean_DDD_cols}
\title{Clean DDD columns}
\usage{
clean_DDD_cols(ATCtable)
}
\arguments{
\item{ATCtable}{An ATC table with unsafe column names.}
}
\value{
An ATC table with safe column names.
}
\description{
pivot_wider() of DDD cols creates unsafe column names. Rename those columns.
}
