% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_atc_html_table.R
\name{extract_atc_ddd_tbl}
\alias{extract_atc_ddd_tbl}
\title{Extract a L5 ATC/DDD table from a page object.}
\usage{
extract_atc_ddd_tbl(atcpageobj)
}
\arguments{
\item{atcpageobj}{An ATC page object: a list of the atc_code, page URL, and the page contents.}
}
\value{
A L5 ATC/DDD table.
}
\description{
The ATC/DDD index is reported in an HTML table; get the table.
}
