% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_atc_hierarchy.R
\name{extract_atc_hierarchy}
\alias{extract_atc_hierarchy}
\title{Extract the ATC hierarchy from an ACT page object}
\usage{
extract_atc_hierarchy(atcpageobj, atc_type)
}
\arguments{
\item{atcpageobj}{An ATC page object: a list of the atc_code, page URL, and the page contents.}

\item{atc_type}{String: an ACT type ('atc'|'atcvet').}
}
\value{
A list of page data.
}
\description{
Extract the ATC hierarchy from an ACT page object
}
