% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_codes.R
\name{get_child_codes}
\alias{get_child_codes}
\title{Get child codes}
\usage{
get_child_codes(pagedata)
}
\arguments{
\item{pagedata}{Page data.}
}
\value{
Child codes.
}
\description{
Get child codes
}
