% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polite_scrape_ATC_page.R
\name{polite_scrape_ATC_code}
\alias{polite_scrape_ATC_code}
\title{Politely scrape an ATC index item page}
\usage{
polite_scrape_ATC_code(atc_code, baseURL = "")
}
\arguments{
\item{atc_code}{String: an ATC index code.}

\item{baseURL}{String: the base URL to prepend to the ATC index code.}
}
\value{
An ATC page object: a list of the atc_code, page URL, and the page contents.
}
\description{
Politely scrape an ATC index item page
}
\details{
The baseURL for the ATC index: "https://atcddd.fhi.no/atc_ddd_index/?code="
The baseURL for the ATCvet index: "https://www.whocc.no/atcvet/atcvet_index/?code="
}
