% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ATC_page.R
\name{save_atc_page}
\alias{save_atc_page}
\title{Save an ATC index item page}
\usage{
save_atc_page(pageobj, atc_type = "")
}
\arguments{
\item{pageobj}{An ATC page object: a list of the atc_code, page URL, and the page contents.}

\item{atc_type}{String: an ACT type ('atc'|'atcvet').}
}
\value{
NA
}
\description{
Save an ATC index item page
}
