% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ATC_scrape.R
\name{save_atc_scrape}
\alias{save_atc_scrape}
\title{Save scrape session}
\usage{
save_atc_scrape(atc_L1_code, atc_type)
}
\arguments{
\item{atc_L1_code}{String: an ATC level 01 group code.}

\item{atc_type}{String: an ACT type ('atc'|'atcvet').}
}
\value{
None.
}
\description{
Save all objects in the global environment.
}
