% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_atc_table.R
\name{write_atc_table_csv}
\alias{write_atc_table_csv}
\title{Write an ATC table to .csv}
\usage{
write_atc_table_csv(atc_table, atc_L1_code = "", atc_type = "")
}
\arguments{
\item{atc_table}{An ATC table.}

\item{atc_L1_code}{String: an ATC level 01 group code.}

\item{atc_type}{String: an ACT type ('ATC'|'ATCvet').}
}
\value{
None.
}
\description{
Write an ATC table to .csv
}
